Prevent position counter from overflow when decoding waypoints. Fixes crashes on levels with thousands of waypoints.
git-svn-id: https://freedroid.svn.sourceforge.net/svnroot/freedroid@4560 524c2bfb-fa3f-0410-b5a2-92f4e1dc5b61

+++ src/map.c
@@ -775,7 +775,7 @@ static char *decode_waypoints(level *loadlevel, char *data)
 	if ((wp_end = strstr(data, WP_END_STRING)) == NULL)
 		return NULL;
 
-	short int curlinepos = 0;
+	int curlinepos = 0;
 	this_line = (char *)MyMalloc(4096);
 	
 	while (1) {
